chore: abstract code into hook
